[EG-1209] Fix menu: check for missing menu entries of existing pages
Menu entries added for multiple files in Corda Enterprise 4.5 and Corda OS 4.5 - see ticket for full details of menus/entries added. Note that in some cases, adding these menu options involved changing menu weights in additional files that are outside of the scope of the ticket, to ensure that menu entries were listed in the correct position in the navigation structure.