[EG-2080] [Docs] Update Corda OS 4.4 and 4.5 CorDapp quickstart and dev pages to reflect new samples structure on github
Multiple updates to ensure that there is only one topic in the OS and Enterprise v4.5 content that references how to set up your development environment and one topic describing how to run an example CorDapp. Removed the "Getting started developing CorDapps" section from the docs menus. Deleted the obsolete Corda Samples page. Full details of the changes made are on the ticket.