[EG-2239/EG-3347] Re-order and edit documentation tutorials
Re-ordered and edited documentation tutorials.

The majority of work is in the tutorials files. However, there were a number of cross-references in other non-tutorial files that had to be updated as a result of re-naming some of the files. These files therefore have to also be included in this PR. In some instances, some minor, non-controversial text edits were also made when updating links in the non-tutorial files.